refactor: migrate components to use auto-selected factories
Update all component classes to import from _ComponentFactoriesAuto:

- HC_constant_elastance.py
- HC_mixed_elastance.py
- HC_mixed_elastance_pp.py
- R_component.py
- Rc_component.py
- Rlc_component.py
- Valve_maynard.py
- Valve_non_ideal.py
- Valve_simple_bernoulli.py

Change: from ._ComponentFactories → from ._ComponentFactoriesAuto

This allows automatic selection between:
- _ComponentFactoriesOptimized (when Cython available)
- _ComponentFactories (fallback to Numba)

No functional changes - maintains identical behavior.
