+++ src/ModularCirc/Components/HC_mixed_elastance.py
@@ -1,5 +1,5 @@
 from .ComponentBase import ComponentBase
-from ._ComponentFactories import ComponentFunctionFactory, ElastanceFactory
+from ._ComponentFactoriesAuto import ComponentFunctionFactory, ElastanceFactory
 from ..HelperRoutines import activation_function_1, chamber_volume_rate_change
 from ..Time import TimeClass
