+++ src/ModularCirc/Components/Valve_maynard.py
@@ -1,5 +1,5 @@
 from .ComponentBase import ComponentBase
-from ._ComponentFactories import ComponentFunctionFactory 
+from ._ComponentFactoriesAuto import ComponentFunctionFactory 
 from ..Time import TimeClass
 from ..StateVariable import StateVariable
